fix logic used to find CONDA_ROOT
If not in the base/root environment or a named environment use the
CONDA_EXE variable to call conda and determine the root prefix.
If this variable is not set, fall-back to a 'conda' command which
may by on PATH.

When not on Windows set the creationflags parameter to 0.

Rename the CONDA_PREFIX variable to ENV_PREFIX which better represents
the path.

fixes #643